+++ wrapper/src/api/providers/benchmark/benchmarkProvider/intersection/index.ts
@@ -1,0 +1,4 @@
+import { IntersectionClusters } from './clusters';
+export { IntersectionCache } from './cache';
+
+export class Intersection extends IntersectionClusters {}
